refactor(workbench): split ConnectedWorkbench into WorkbenchProviders + Workbench
