+++ packages/workbench/src/index.ts
@@ -1,4 +1,5 @@
-export * from './ConnectedWorkbench';
 export * from './debug';
 export * from './runtime';
 export * from './tayori';
+export * from './Workbench';
+export * from './WorkbenchProviders';
